Add Experience and Level Up Notification Text and refactor XPChangedDelegate to contain the XP received

+++ Source/SurvivalCraft/Private/UI/WidgetController/SCInventoryMenuWidgetController.cpp
@@ -25,7 +25,7 @@ void USCInventoryMenuWidgetController::BindCallbacksToDependencies()
 	GetSCPS()->OnLevelChangedDelegate.AddLambda(
 	[this](int32 NewLevel)
 	{
-		OnPlayerLevelChangedDelegate.Broadcast(NewLevel);	
+		OnPlayerLevelChangedDelegate.Broadcast(static_cast<float>(NewLevel));	
 	}
 );
 }
